+++ xclim/sdba/utils.py
@@ -456,7 +456,7 @@ def interp_on_quantiles(
         output_core_dims=[[dim]],
         vectorize=True,
         dask="parallelized",
-        output_dtypes=[np.float],
+        output_dtypes=[yq.dtype],
     )